+++ client/run_group.go
@@ -8,7 +8,7 @@ import (
 
 // Group is used to group a list of clients and start/stop them
 // currently a thin wrapper around run.Group that adds a Stop() function
-type Group struct {
+type RunGroup struct {
 	name     string
 	stop     chan struct{}
 	stopOnce sync.Once
@@ -16,18 +16,18 @@ type Group struct {
 }
 
 // NewGroup creates a new client group
-func NewGroup(name string) *Group {
-	return &Group{name: name, stop: make(chan struct{})}
+func NewRunGroup(name string) *RunGroup {
+	return &RunGroup{name: name, stop: make(chan struct{})}
 }
 
 // Add client to group
-func (g *Group) Add(client RunStop) {
+func (g *RunGroup) Add(client RunStop) {
 	g.group.Add(client.Run, client.Stop)
 }
 
 // Run clients. This function blocks until error or stopped.
 // all clients must be added before runner is started
-func (g *Group) Run() error {
+func (g *RunGroup) Run() error {
 	g.group.Add(func() error {
 		<-g.stop
 		return nil
@@ -41,6 +41,6 @@ func (g *Group) Run() error {
 }
 
 // Stop clients
-func (g *Group) Stop(_ error) {
+func (g *RunGroup) Stop(_ error) {
 	g.stopOnce.Do(func() { close(g.stop) })
 }